Change serverlist format to JSON
In preparation of upcoming changes the format of the serverlist file is
changed to JSON because the old format of colon separated values makes
it unnecessarily hard to add additional fields as you have to specify
each field before it (even if the default value would be fine).
The new script convert_serverlist.php takes care of converting your old
server.txt to the new format. After executing it (check permissions on
your web directory if you want to execute it through the web browser)
you have to adjust your config.php to point to the newly created JSON
file.