Use python2.6 in Updian's virtual environment if available
Using python 2.7 on debian wheezy we encountered high CPU usage upon
connecting. It seems to be related to paramiko hashing already hashed
entries in the ssh known_hosts file (see
<https://github.com/paramiko/paramiko/issues/191>). The patch proposed
on github fixes the issue but because we can't really ask every Updian
user to patch paramiko we are using /usr/bin/python2.6 when it is
available which also alleviates the situation.